feat: Add new flag to run ALL preprocessing steps
This commit adds a new flag '-P' to force the execution of all preprocessing steps. If the flag is set, all other flags will be ignored. The download.py script is now set to exit if it fails to download the dataset. The extract_bboxes.py, generate_crops.py, and separate_crops.py scripts are now set to exit if they fail.